# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/deployments": {
            "get": {
                "description": "Retrieves a list of deployments from the Kubernetes cluster, optionally filtered by namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "List deployments",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Maximum number of deployments to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination token for fetching more deployments",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter deployments by",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of deployments",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing deployment in the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Update an existing deployment",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deployment request body",
                        "name": "deployment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeploymentUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated the deployment",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new deployment in the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Create a new deployment",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Deployment request body",
                        "name": "deployment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeploymentCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Suxccessfully created deployment",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or error message",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
        "/deployments/{id}": {
            "get": {
                "description": "Retrieves a deployment from the Kubernetes cluster by its name or UID, optionally filtered by namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Get a deployment by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter the deployment by",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the deployment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested deployment",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a deployment from the Kubernetes cluster by its name or UID, optionally filtered by namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "summary": "Delete a deployment by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter the deployment by",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the deployment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Bad request or error message",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.FailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "model.Deployment": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "description": "Standard object's metadata.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ObjectMeta"
                        }
                    ]
                },
                "spec": {
                    "description": "Specification of the desired behavior of the Deployment.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.DeploymentSpec"
                        }
                    ]
                },
                "status": {
                    "description": "Most recently observed status of the Deployment.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.DeploymentStatus"
                        }
                    ]
                }
            }
        },
        "model.DeploymentCondition": {
            "type": "object",
            "properties": {
                "lastTransitionTime": {
                    "description": "Last time the condition transitioned from one status to another.",
                    "type": "string"
                },
                "lastUpdateTime": {
                    "description": "The last time this condition was updated.",
                    "type": "string"
                },
                "message": {
                    "description": "A human readable message indicating details about the transition.",
                    "type": "string"
                },
                "reason": {
                    "description": "The reason for the condition's last transition.",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the condition, one of True, False, Unknown.",
                    "type": "string"
                },
                "type": {
                    "description": "Type of deployment condition.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.DeploymentConditionType"
                        }
                    ]
                }
            }
        },
        "model.DeploymentConditionType": {
            "type": "string",
            "enum": [
                "Available",
                "Progressing",
                "ReplicaFailure"
            ],
            "x-enum-varnames": [
                "DeploymentAvailable",
                "DeploymentProgressing",
                "DeploymentReplicaFailure"
            ]
        },
        "model.DeploymentCreateRequest": {
            "type": "object",
            "properties": {
                "deployment": {
                    "$ref": "#/definitions/model.Deployment"
                },
                "opts": {
                    "$ref": "#/definitions/model.CreateOptions"
                }
            }
        },
        "model.DeploymentSpec": {
            "type": "object",
            "properties": {
                "minReadySeconds": {
                    "description": "Minimum number of seconds for which a newly created pod should be ready\nwithout any of its container crashing, for it to be considered available.\nDefaults to 0 (pod will be considered available as soon as it is ready)\n+optional",
                    "type": "integer"
                },
                "paused": {
                    "description": "Indicates that the deployment is paused.\n+optional",
                    "type": "boolean"
                },
                "progressDeadlineSeconds": {
                    "description": "The maximum time in seconds for a deployment to make progress before it\nis considered to be failed. The deployment controller will continue to\nprocess failed deployments and a condition with a ProgressDeadlineExceeded\nreason will be surfaced in the deployment status. Note that progress will\nnot be estimated during the time a deployment is paused. Defaults to 600s.",
                    "type": "integer"
                },
                "replicas": {
                    "description": "Number of desired pods. This is a pointer to distinguish between explicit\nzero and not specified. Defaults to 1.\n+optional",
                    "type": "integer"
                },
                "revisionHistoryLimit": {
                    "description": "The number of old ReplicaSets to retain to allow rollback.\nThis is a pointer to distinguish between explicit zero and not specified.\nDefaults to 10.\n+optional",
                    "type": "integer"
                },
                "selector": {
                    "description": "Label selector for pods. Existing ReplicaSets whose pods are\nselected by this will be the ones affected by this deployment.\nIt must match the pod template's labels.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.LabelSelector"
                        }
                    ]
                },
                "strategy": {
                    "description": "The deployment strategy to use to replace existing pods with new ones.\n+optional\ns",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.DeploymentStrategy"
                        }
                    ]
                },
                "template": {
                    "description": "Template describes the pods that will be created.\nThe only allowed template.spec.restartPolicy value is \"Always\".",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PodTemplateSpec"
                        }
                    ]
                }
            }
        },
        "model.DeploymentStatus": {
            "type": "object",
            "properties": {
                "availableReplicas": {
                    "description": "Total number of available pods (ready for at least minReadySeconds) targeted by this deployment.\n+optional",
                    "type": "integer"
                },
                "collisionCount": {
                    "description": "Count of hash collisions for the Deployment. The Deployment controller uses this\nfield as a collision avoidance mechanism when it needs to create the name for the\nnewest ReplicaSet.\n+optional",
                    "type": "integer"
                },
                "conditions": {
                    "description": "Represents the latest available observations of a deployment's current state.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DeploymentCondition"
                    }
                },
                "observedGeneration": {
                    "description": "The generation observed by the deployment controller.\n+optional",
                    "type": "integer"
                },
                "readyReplicas": {
                    "description": "readyReplicas is the number of pods targeted by this Deployment with a Ready Condition.\n+optional",
                    "type": "integer"
                },
                "replicas": {
                    "description": "Total number of non-terminated pods targeted by this deployment (their labels match the selector).\n+optional",
                    "type": "integer"
                },
                "unavailableReplicas": {
                    "description": "Total number of unavailable pods targeted by this deployment. This is the total number of\npods that are still required for the deployment to have 100% available capacity. They may\neither be pods that are running but not yet available or pods that still have not been created.\n+optional",
                    "type": "integer"
                },
                "updatedReplicas": {
                    "description": "Total number of non-terminated pods targeted by this deployment that have the desired template spec.\n+optional",
                    "type": "integer"
                }
            }
        },
        "model.DeploymentStrategy": {
            "type": "object",
            "properties": {
                "type": {
                    "description": "Rolling update config params. Present only if DeploymentStrategyType =\nRollingUpdate.\n---\nTODO: Update this to follow our convention for oneOf, whatever we decide it\nto be.\n+optional\nRollingUpdate *RollingUpdateDeployment `json:\"rollingUpdate,omitempty\"`\nType of deployment. Can be \"Recreate\" or \"RollingUpdate\". Default is RollingUpdate.\n+optional",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.DeploymentStrategyType"
                        }
                    ]
                }
            }
        },
        "model.DeploymentStrategyType": {
            "type": "string",
            "enum": [
                "Recreate",
                "RollingUpdate"
            ],
            "x-enum-varnames": [
                "RecreateDeploymentStrategyType",
                "RollingUpdateDeploymentStrategyType"
            ]
        },
        "model.DeploymentUpdateRequest": {
            "type": "object",
            "properties": {
                "deployment": {
                    "$ref": "#/definitions/model.Deployment"
                },
                "opts": {
                    "$ref": "#/definitions/model.UpdateOptions"
                }
            }
        },
...
        "model.LabelSelector": {
            "type": "object",
            "properties": {
                "matchExpressions": {
                    "description": "matchExpressions is a list of label selector requirements. The requirements are ANDed.\n+optional\n+listType=atomic",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.LabelSelectorRequirement"
                    }
                },
                "matchLabels": {
                    "description": "matchLabels is a map of {key,value} pairs. A single {key,value} in the matchLabels\nmap is equivalent to an element of matchExpressions, whose key field is \"key\", the\noperator is \"In\", and the values array contains only \"value\". The requirements are ANDed.\n+optional",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.LabelSelectorRequirement": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "key is the label key that the selector applies to.",
                    "type": "string"
                },
                "operator": {
                    "description": "operator represents a key's relationship to a set of values.\nValid operators are In, NotIn, Exists and DoesNotExist.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.LabelSelectorOperator"
                        }
                    ]
                },
                "values": {
                    "description": "values is an array of string values. If the operator is In or NotIn,\nthe values array must be non-empty. If the operator is Exists or DoesNotExist,\nthe values array must be empty. This array is replaced during a strategic\nmerge patch.\n+optional\n+listType=atomic",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.ObjectMeta": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "creationTimestamp": {
                    "type": "string"
                },
                "deletionGracePeriodSeconds": {
                    "type": "integer"
                },
                "deletionTimestamp": {
                    "type": "string"
                },
                "finalizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "generateName": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ownerReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OwnerReference"
                    }
                },
                "resourceVersion": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.OwnerReference": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "blockOwnerDeletion": {
                    "type": "boolean"
                },
                "controller": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.PodSpec": {
            "type": "object",
            "properties": {
                "activeDeadlineSeconds": {
                    "type": "integer"
                },
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Container"
                    }
                },
                "initContainers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Container"
                    }
                },
                "terminationGracePeriodSeconds": {
                    "type": "integer"
                },
                "tolerations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Toleration"
                    }
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Volume"
                    }
                }
            }
        },
...
        "model.PodTemplateSpec": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/model.ObjectMeta"
                },
                "spec": {
                    "$ref": "#/definitions/model.PodSpec"
                }
            }
        },
...
        "model.UpdateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
        "model.Volume": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}